feat: restructure async email queue into dedicated jobs and services
